backend/requirements.txt
flask
Jinja2
mysql-connector-python
